# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Hoye Secondary School Online Admissions API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface ApplicationSubmitResult {
  /** Tracking reference number assigned to this application */
  refNumber: string;
  message: string;
}

export type ApplicationStatusStatus = typeof ApplicationStatusStatus[keyof typeof ApplicationStatusStatus];


export const ApplicationStatusStatus = {
  pending: 'pending',
  under_review: 'under_review',
  accepted: 'accepted',
  rejected: 'rejected',
  waitlisted: 'waitlisted',
} as const;

export interface ApplicationStatus {
  refNumber: string;
  firstName: string;
  lastName: string;
  gradeApplying: string;
  status: ApplicationStatusStatus;
  submittedAt: string;
  updatedAt: string;
}

export interface RecoverResult {
  refNumber: string;
}

export type ApplicationInputGender = typeof ApplicationInputGender[keyof typeof ApplicationInputGender];


export const ApplicationInputGender = {
  Male: 'Male',
  Female: 'Female',
} as const;

export interface ApplicationInput {
  firstName: string;
  lastName: string;
  /** @nullable */
  middleName?: string | null;
  /** @nullable */
  preferredName?: string | null;
  idNumber: string;
  /** @nullable */
  birthCertNumber?: string | null;
  dob: string;
  /** @nullable */
  age?: number | null;
  gender: ApplicationInputGender;
  nationality: string;
  /** @nullable */
  citizenship?: string | null;
  /** @nullable */
  countryOfBirth?: string | null;
  homeLanguage: string;
  /** @nullable */
  secondLanguage?: string | null;
  /** @nullable */
  learningLanguage?: string | null;
  /** @nullable */
  religion?: string | null;
  /** @nullable */
  mobileNumber?: string | null;
  /** @nullable */
  email?: string | null;
  residentialAddress: string;
  city: string;
  province: string;
  /** @nullable */
  postalCode?: string | null;
  country: string;
  /** @nullable */
  livesWith?: string | null;
  /** @nullable */
  totalSiblings?: number | null;
  /** @nullable */
  familyPosition?: string | null;
  /** @nullable */
  siblingsAtSchool?: string | null;
  /** @nullable */
  preferredStartYear?: number | null;
  /** @nullable */
  gradePassed?: string | null;
  /** @nullable */
  yearPassedHighest?: number | null;
  /** @nullable */
  currentGrade?: string | null;
  /** @nullable */
  currentAcademicYear?: number | null;
  prevSchoolName: string;
  /** @nullable */
  prevSchoolAddress?: string | null;
  /** @nullable */
  prevSchoolPhone?: string | null;
  /** @nullable */
  reasonLeaving?: string | null;
  /** @nullable */
  averagePercentage?: string | null;
  /** @nullable */
  bestSubject?: string | null;
  /** @nullable */
  weakestSubject?: string | null;
  /** @nullable */
  numSubjectsPassed?: number | null;
  gradeApplying: string;
  /** @nullable */
  chosenStream?: string | null;
  /** @nullable */
  optionalSubject?: string | null;
  /** @nullable */
  needsSupport?: string | null;
  /** @nullable */
  needsExtraLessons?: string | null;
  /** @nullable */
  achievements?: string | null;
  /** @nullable */
  sportsParticipation?: string | null;
  /** @nullable */
  leadershipRoles?: string | null;
  /** @nullable */
  extracurricular?: string | null;
  /** @nullable */
  hasDisciplineHistory?: string | null;
  /** @nullable */
  disciplineDetails?: string | null;
  motivationToJoin: string;
  referralSource: string;
  /** @nullable */
  hasTeacherRelative?: string | null;
  /** @nullable */
  teacherName?: string | null;
  /** @nullable */
  teacherSurname?: string | null;
  /** @nullable */
  teacherPhone?: string | null;
  /** @nullable */
  teacherRelationship?: string | null;
  /** @nullable */
  hasMedicalAid?: string | null;
  /** @nullable */
  medAidProvider?: string | null;
  /** @nullable */
  medAidNumber?: string | null;
  /** @nullable */
  hasFamilyDoctor?: string | null;
  /** @nullable */
  doctorName?: string | null;
  /** @nullable */
  doctorPhone?: string | null;
  emergencyName: string;
  emergencyPhone: string;
  /** @nullable */
  emergencyRelationship?: string | null;
  /** @nullable */
  medicalAllergies?: string | null;
  /** @nullable */
  medicalConditions?: string | null;
  /** @nullable */
  currentMedication?: string | null;
  /** @nullable */
  medAsthma?: string | null;
  /** @nullable */
  medEpilepsy?: string | null;
  /** @nullable */
  medDiabetes?: string | null;
  /** @nullable */
  hasDisability?: string | null;
  /** @nullable */
  disabilityDetails?: string | null;
  /** @nullable */
  needsCounselling?: string | null;
  /** @nullable */
  sportsAllowed?: string | null;
  /** @nullable */
  allowEmergencyTreatment?: string | null;
  guardianFirstName: string;
  guardianLastName: string;
  guardianId: string;
  guardianRelationship: string;
  /** @nullable */
  guardianMaritalStatus?: string | null;
  guardianPhone: string;
  /** @nullable */
  guardianAltPhone?: string | null;
  /** @nullable */
  guardianWhatsapp?: string | null;
  /** @nullable */
  guardianEmail?: string | null;
  /** @nullable */
  guardianEmploymentStatus?: string | null;
  /** @nullable */
  guardianOccupation?: string | null;
  /** @nullable */
  guardianEmployer?: string | null;
  /** @nullable */
  guardianIncomeRange?: string | null;
  /** @nullable */
  guardianHomeAddress?: string | null;
  /** @nullable */
  preferredCommunication?: string | null;
  /** @nullable */
  hasSecondGuardian?: string | null;
  /** @nullable */
  guardian2Name?: string | null;
  /** @nullable */
  guardian2Phone?: string | null;
  /** @nullable */
  permissionPhotos?: string | null;
  confirmTruth: string;
  agreePolicies: string;
  digitalSignature: string;
  submissionDate: string;
  /** @nullable */
  additionalNotes?: string | null;
}

export type RecoverReferenceParams = {
/**
 * Learner ID or passport number
 */
idNumber: string;
/**
 * Parent/guardian phone number
 */
phone: string;
};