# lib/api-zod/src/generated/types/applicationStatus.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Hoye Secondary School Online Admissions API
 * OpenAPI spec version: 0.1.0
 */
import type { ApplicationStatusStatus } from './applicationStatusStatus';

export interface ApplicationStatus {
  refNumber: string;
  firstName: string;
  lastName: string;
  gradeApplying: string;
  status: ApplicationStatusStatus;
  submittedAt: Date;
  updatedAt: Date;
}