# lib/api-zod/src/generated/types/errorResponse.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Hoye Secondary School Online Admissions API
 * OpenAPI spec version: 0.1.0
 */

export interface ErrorResponse {
  error: string;
}